fix: move vega utils out of canvas utils
