+++ web-common/src/features/templates/charts/stacked-area.ts
@@ -1,11 +1,11 @@
+import {
+  sanitizeValueForVega,
+  sanitizeValuesForSpec,
+} from "@rilldata/web-common/components/vega/util";
 import type { TooltipValue } from "@rilldata/web-common/features/canvas/components/charts/types";
 import { ScrubBoxColor } from "@rilldata/web-common/features/dashboards/time-series/chart-colors";
 import type { ChartField } from "./build-template";
-import {
-  multiLayerBaseSpec,
-  sanitizeValueForVega,
-  sanitizeValuesForSpec,
-} from "./utils";
+import { multiLayerBaseSpec } from "./utils";
 
 export function buildStackedArea(
   timeField: ChartField,
